cleaned out old tests. updated code. everything is building, no tests yet.
