--- core/src/test/scala/org/apache/spark/sql/streaming/eventhubs/EventHubsOffsetSuite.scala
@@ -1,39 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.sql.streaming.eventhubs
-
-import org.apache.spark.eventhubs.common.NameAndPartition
-import org.apache.spark.sql.streaming.OffsetSuite
-import org.apache.spark.sql.test.SharedSQLContext
-
-class EventHubsOffsetSuite extends OffsetSuite with SharedSQLContext {
-
-  val eventHubsBatchRecord = EventHubsSourceOffset(0L,
-                                                   Map(NameAndPartition("eventhub", 0) -> 0L,
-                                                       NameAndPartition("eventhub", 1) -> 100L,
-                                                       NameAndPartition("eventhub", 2) -> 200L))
-
-  test("basic serialization and deserialization of Eventhubs batch record") {
-
-    val deserializedEventhubsBatchRecord = JsonUtils.partitionAndSeqNum(eventHubsBatchRecord.json)
-
-    assert(deserializedEventhubsBatchRecord.batchId === eventHubsBatchRecord.batchId)
-    assert(
-      deserializedEventhubsBatchRecord.partitionToSeqNos === eventHubsBatchRecord.partitionToSeqNos)
-  }
-}
